sleep styles added to seperate file
