+++ Aura_frontend/app/(tabs)/SleepScreen/SleepBetterScreen.tsx
@@ -3,17 +3,17 @@ import {
   View,
   Text,
   TouchableOpacity,
-  StyleSheet,
   StatusBar,
   ScrollView,
 } from "react-native";
+import styles from "./SleepBetterScreen.styles";
 import dayjs from "dayjs";
 import { router } from "expo-router";
 import { SafeAreaView } from "react-native-safe-area-context";
-import BackButton from "../components/BackButton";
+import BackButton from "../../components/BackButton";
 import { Feather } from "@expo/vector-icons";
-import SleepChart from "../components/SleepChart";
-import { fetchSleepData } from "../services/sleepService";
+import SleepChart from "../../components/SleepChart";
+import { fetchSleepData } from "../../services/sleepService";
 
 const SleepBetterScreen = () => {
   type sleepRecord = {
@@ -103,7 +103,7 @@ const SleepBetterScreen = () => {
         {/* Time Selector */}
         <View style={styles.timeSelector}>
           <TouchableOpacity
-            onPress={() => router.push("/(tabs)/SleepTimerScreen")}
+            onPress={() => router.push("/(tabs)/SleepScreen/SleepTimerScreen")}
             style={styles.timeSelectorButton}
           >
             <Text style={styles.timeSelectorText}>{getTimeRange()}</Text>
@@ -170,107 +170,5 @@ const SleepBetterScreen = () => {
   );
 };
 
-const styles = StyleSheet.create({
-  container: { 
-    flex: 1, 
-    backgroundColor: "#ffffff" 
-  },
-  scrollContent: { 
-    flexGrow: 1, 
-    paddingBottom: 100 
-  },
-  titleText: {
-    fontSize: 18,
-    fontWeight: "regular",
-  },
-  titleContainer: {
-    alignItems: "center",
-    justifyContent: "space-between",
-    paddingHorizontal: 15,
-    paddingVertical: 15,
-  },
-  outerContainer: {
-    flexDirection: "row",
-    borderColor: "#224831", // green border
-    borderWidth: 1,
-    borderRadius: 12,
-    padding: 10,
-    justifyContent: "space-between",
-    margin: 20,
-  },
-  dateChip: {
-    alignItems: "center",
-    padding: 8,
-    borderRadius: 8,
-    borderWidth: 1,
-    borderColor: "#A5D6A7", // light green
-    marginHorizontal: 4,
-    backgroundColor: "white",
-  },
-  selectedChip: {
-    backgroundColor: "#4CAF50", // selected green
-    borderColor: "#4CAF50",
-  },
-  dayText: {
-    fontSize: 12,
-    fontWeight: "500",
-    color: "#224831",
-  },
-  dateText: {
-    fontSize: 14,
-    fontWeight: "700",
-    color: "#224831",
-  },
-  selectedText: {
-    color: "white",
-  },
-  timeSelector: {
-    flexDirection: "row",
-    borderColor: "224831",
-    borderWidth: 1,
-    borderRadius: 8,
-    padding: 10,
-    marginHorizontal: 100,
-    marginTop: 20,
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  timeSelectorText: {
-    fontWeight: "bold",
-    fontSize: 15,
-    alignItems: "center",
-    paddingHorizontal: 10,
-    textAlign: "center",
-  },
-  timeSelectorButton: {
-    flexDirection: "row",
-    alignItems: "center",
-  },
-  sleepChartContainer: {
-    // marginTop: 10,
-    padding: 10,
-  },
-  sleepChart: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    paddingHorizontal: 20,
-    marginBottom: 10,
-  },
-  weekSelector: { fontWeight: "bold", fontSize: 16, paddingTop: 20 },
-  avgSleepContainer: {
-    // marginTop: 10,
-    paddingHorizontal: 15,
-    alignItems: "center",
-    borderColor: "#4CAF50",
-    borderRadius: 10,
-    borderWidth: 1,
-    padding: 10,
-    marginHorizontal: 30,
-  },
-  avgSleepText: {
-    fontSize: 16,
-    fontWeight: "500",
-  },
-});
 
 export default SleepBetterScreen;
